feat: add nsfw plugin with missav downloader
This commit introduces a new plugin that enables users to download content from MissAV.
The plugin includes a command handler for the `/missav` command, which takes a MissAV link as input.
It then processes the link and downloads the video with the lowest quality using the `missav_dl` function from the `core` module.
A processing message is displayed while the download is in progress.

Signed-off-by: troublescope <tomiprs25@gmail.com>