feat(nsfw.py): implement missav command to download videos
This commit introduces a new command, `/missav`, that allows users
to download videos from MissAV. The command takes a MissAV link as
an argument and prompts the user to select the desired video quality.
The video is then downloaded and sent to the user.

The commit also includes the following changes:

- Added a cache to store links using message.id as key
- Added a callback query handler for quality selection
- Added a function to split videos larger than MAX_SIZE
- Added a function to upload videos using media group
- Added a function to generate thumbnails for videos
- Added a function to get video info (duration)
- Added progress indicator during download and upload
- Added duration to caption

Signed-off-by: troublescope <tomiprs25@gmail.com>